First

+++ HR_System/Models/Setting.cs
@@ -6,8 +6,8 @@ namespace HR_System.Models
     public partial class Setting
     {
         public int SettingId { get; set; }
-        public double? PlusPerhour { get; set; }
-        public double? MinusPerhour { get; set; }
+        public float? PlusPerhour { get; set; }
+        public float? MinusPerhour { get; set; }
         public string? Dayoff1 { get; set; }
         public string? Dayoff2 { get; set; }
     }
